Refactor connect component configs
