+++ kpops/components/streams_bootstrap/producer/producer_app.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from pydantic import BaseConfig, Extra
 from typing_extensions import override
